Refactored island imports according to latest changes
